pyproject.toml
[tool.poetry]
name = "laoflix"
version = "0.1.0"
description = ""
authors = ["xXLAOKOONXx"]

[tool.poetry.dependencies]
python = "3.11"
Pillow = "^9.5.0"
ffmpeg-python = "^0.2.0"
ipykernel = "^6.29.2"
exceptiongroup = "^1.2.0"
mutagen = "^1.47.0"
moviepy = "^1.0.3"
loguru = "^0.7.2"

[tool.poetry.dev-dependencies]
pyinstaller = "^5.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
